Add AppDelegate
